Impute z-score based correlations

Impute z-score from a missing test based on z-scores for other tests, and the genotype matrix

Usage

imputezDecorr(z, X, i, k, lambda = NULL)

Arguments

z

vector of observed z-scores

X

matrix genotype data from a reference panel

i

index of z-scores to impute

k

rank of SVD

lambda

(default: NULL) value used to shrink correlation matrix

Value

data.frame storing:

ID

variant identifier

z.stat

imputed z-statistic

sigSq

variance of imputed z-statistic

r2.pred

metric of accuracy of the imputed z-statistic based on its variance

lambda

correlation shrinkage parameter

Details

Uses implicit covariance and emprical Bayes shrinkage of the eigen-values to accelerate computations. imputez() is cubic in the number of features, p, but imputezDecorr() is the minimum of O(n p^2) and O(n^2p). For large number of features, this is can be a dramatic speedup.

References

Pasaniuc, B., Zaitlen, N., Shi, H., Bhatia, G., Gusev, A., Pickrell, J., ... & Price, A. L. (2014). Fast and accurate imputation of summary statistics enhances evidence of functional enrichment. Bioinformatics, 30(20), 2906-2914.

See also