Impute many z-statistics

Impute many z-statistics given observed z-statistics and data.table of LD.

Usage

run_imputez(z, LDinfo, IDs, maxWindowSize = 200, quiet = FALSE)

Arguments

z

vector of observed z-statistics with unobserved storing value NA

LDinfo

list with data.table storing LD information, and GRanges with position info

IDs

variant names in z to impute

maxWindowSize

max window size around target variant to keep

quiet

default FALSE. If TRUE, suppress progress bar

Value

data.frame storing imputed results:

ID

variant identifier

z.stat

imputed z-statistic

sigSq

variance of imputed z-statistic

r2.pred

metric of accuracy of the imputed z-statistic based on its variance

width

number of SNPs in LD panel used in imputation

Details

Implements method by Pasaniuc, et al. (2014).

References

Pasaniuc B, Zaitlen N, Shi H, Bhatia G, Gusev A, Pickrell J, Hirschhorn J, Strachan DP, Patterson N, Price AL (2014). “Fast and accurate imputation of summary statistics enhances evidence of functional enrichment.” Bioinformatics, 30(20), 2906--2914.

See also