Impute z-score based correlations

Impute z-score from a missing test based on z-scores for other tests, and the correlation matrix between z-scores

Usage

imputez(z, Sigma, i, lambda = 0.1)

Arguments

z

vector of observed z-scores

Sigma

matrix of correlation between z-scores

i

index of z-scores to impute

lambda

value used to shrink correlation matrix

Value

data.frame storing:

ID

variant identifier

z.stat

imputed z-statistic

sigSq

variance of imputed z-statistic

r2.pred

metric of accuracy of the imputed z-statistic based on its variance

Details

Implements method by Pasaniuc, et al. (2014).

References

Pasaniuc, B., Zaitlen, N., Shi, H., Bhatia, G., Gusev, A., Pickrell, J., ... & Price, A. L. (2014). Fast and accurate imputation of summary statistics enhances evidence of functional enrichment. Bioinformatics, 30(20), 2906-2914.